Abstract out some common code
